Update Guidebooks in README
We now have a few new guidebooks available on the WordPress VIP documentation site:
- https://docs.wpvip.com/guidebooks/get-started/
- https://docs.wpvip.com/guidebooks/develop-on-wpvip/
- https://docs.wpvip.com/guidebooks/prepare-for-launch/

As these guidebooks contain the concepts shown by the current links as well as additional essential information about developing on WordPress VIP, @enfysbook recommended that we add these guidebooks to the README of the VIP Skeleton. 

I agree and have proposed the following edit.